refactor: axios 리펙토링 및 resultData에 대한 코드도 가독성과 깔끔하게 관리하기위해 util.js파일을 만들고 resultData 리펙토링
